+++ src/lib/components/AgGrid.react.js
@@ -12,7 +12,7 @@ function getGrid(enable) {
 /**
  * Dash interface to AG Grid, a powerful tabular data component.
  */
-export default class DashAgGrid extends Component {
+class DashAgGrid extends Component {
     constructor(props) {
         super(props);
 
@@ -23,6 +23,8 @@ export default class DashAgGrid extends Component {
 
         this.buildArray = this.buildArray.bind(this);
     }
+
+    static dashRenderType = true;
 
     buildArray(arr1, arr2) {
         if (arr1) {
@@ -750,10 +752,17 @@ DashAgGrid.propTypes = {
      * Other ag-grid options
      */
     dashGridOptions: PropTypes.object,
+
+    /**
+     * dashRenderType to determine why grid is rendering
+     */
+    dashRenderType: PropTypes.string,
 };
 
 export const propTypes = DashAgGrid.propTypes;
 export const defaultProps = DashAgGrid.defaultProps;
+
+export default DashAgGrid;
 
 export const apiGetters = {};