+++ CHANGELOG.md
@@ -4,6 +4,16 @@ All notable changes to `dash-ag-grid` will be documented in this file.
 This project adheres to [Semantic Versioning](https://semver.org/).
 Links "DE#nnn" prior to version 2.0 point to the Dash Enterprise closed-source Dash AG Grid repo
 
+## [unreleased]
+
+### Fixed
+- [#328](https://github.com/plotly/dash-ag-grid/pull/328) fixed issue where `getRowStyle` wast able to be passed as a complete function
+
+### Changed
+- [#328](https://github.com/plotly/dash-ag-grid/pull/328) 
+  - bumping to v`32.3` for the grid
+  - the grid checks if updates are from an internal source (3.0.2 dash) and will selectively rerender if so
+
 ## [31.3.1] - 2025-03-17
 
 ### Fixed